requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
azure-core==1.26.3
azure-cosmos==4.3.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
DateTime==4.9
frozenlist==1.3.3
httplib2==0.20.4
icalendar==4.0.9
icalevents==0.1.27
idna==3.4
inflect==6.0.2
iso8601==1.1.0
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.logging==3.1.2
jaraco.stream==3.0.3
jaraco.text==3.11.1
more-itertools==9.0.0
multidict==6.0.4
openai==0.27.0
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
regex==2022.10.31
requests==2.28.2
six==1.16.0
tempora==5.2.1
tqdm==4.64.1
twitchio==2.5.0
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
zope.interface==5.5.2
